Pull

+++ module/Model/Package/PackageManager.php
@@ -1,8 +1,9 @@
 <?php
+
 /**
  * Package manager
  *
- * Copyright (C) 2011-2015 Holger Schletz <holger.schletz@web.de>
+ * Copyright (C) 2011-2022 Holger Schletz <holger.schletz@web.de>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the Free
@@ -21,7 +22,7 @@
 
 namespace Model\Package;
 
-use Zend\Db\Sql\Predicate;
+use Laminas\Db\Sql\Predicate;
 
 /**
  * Package manager
@@ -30,16 +31,16 @@ class PackageManager
 {
     /**
      * Service manager
-     * @var \Zend\ServiceManager\ServiceManager
+     * @var \Laminas\ServiceManager\ServiceManager
      */
     protected $_serviceManager;
 
     /**
      * Constructor
      *
-     * @param \Zend\ServiceManager\ServiceManager $serviceManager
+     * @param \Laminas\ServiceManager\ServiceManager $serviceManager
      */
-    public function __construct(\Zend\ServiceManager\ServiceManager $serviceManager)
+    public function __construct(\Laminas\ServiceManager\ServiceManager $serviceManager)
     {
         $this->_serviceManager = $serviceManager;
     }
@@ -83,7 +84,7 @@ class PackageManager
             $package->exchangeArray($storage->readMetadata($package['Id']) + $package->getArrayCopy());
             return $package;
         } catch (\Exception $e) {
-            throw new RuntimeException($e->getMessage(), (integer) $e->getCode(), $e);
+            throw new RuntimeException($e->getMessage(), (int) $e->getCode(), $e);
         }
     }
 
@@ -92,9 +93,9 @@ class PackageManager
      *
      * @param string $order Property to sort by
      * @param string $direction One of [asc|desc]
-     * @return \Zend\Db\ResultSet\AbstractResultSet Result set producing \Model\Package\Package
+     * @return \Laminas\Db\ResultSet\AbstractResultSet Result set producing \Model\Package\Package
      */
-    public function getPackages($order=null, $direction='asc')
+    public function getPackages($order = null, $direction = 'asc')
     {
         $clientConfig = $this->_serviceManager->get('Database\Table\ClientConfig');
         $groupInfo = $this->_serviceManager->get('Database\Table\GroupInfo');
@@ -104,19 +105,19 @@ class PackageManager
         $subquery = $clientConfig->getSql()->select();
         $subquery->columns(array(new Predicate\Literal('COUNT(hardware_id)')))
                  ->where(
-                     array('name' => 'DOWNLOAD', 'ivalue' => new \Zend\Db\Sql\Literal('fileid'))
+                     array('name' => 'DOWNLOAD', 'ivalue' => new \Laminas\Db\Sql\Literal('fileid'))
                  );
 
         $groups = $groupInfo->getSql()->select()->columns(array('hardware_id'));
-        $nonNotified = clone $subquery;
-        $nonNotified->where(new Predicate\IsNull('tvalue'))
-                    ->where(new Predicate\NotIn('hardware_id', $groups));
+        $pending = clone $subquery;
+        $pending->where(new Predicate\IsNull('tvalue'))
+                ->where(new Predicate\NotIn('hardware_id', $groups));
+
+        $running = clone $subquery;
+        $running->where(array('tvalue' => \Model\Package\Assignment::RUNNING));
 
         $success = clone $subquery;
         $success->where(array('tvalue' => \Model\Package\Assignment::SUCCESS));
-
-        $notified = clone $subquery;
-        $notified->where(array('tvalue' => \Model\Package\Assignment::NOTIFIED));
 
         $error = clone $subquery;
         $error->where(new Predicate\Like('tvalue', \Model\Package\Assignment::ERROR_PREFIX . '%'));
@@ -131,9 +132,9 @@ class PackageManager
                 'size',
                 'osname',
                 'comment',
-                'num_nonnotified' => new Predicate\Expression('?', array($nonNotified)),
+                'num_pending' => new Predicate\Expression('?', array($pending)),
+                'num_running' => new Predicate\Expression('?', array($running)),
                 'num_success' => new Predicate\Expression('?', array($success)),
-                'num_notified' => new Predicate\Expression('?', array($notified)),
                 'num_error' => new Predicate\Expression('?', array($error)),
             )
         );
@@ -166,122 +167,10 @@ class PackageManager
      * @param array $data Package data
      * @param bool $deleteSource Delete source file as soon as possible
      * @throws RuntimeException if a package with the requested name already exists or an error occurs
-     * @throws \InvalidArgumentException if 'Platform' key is not a valid value
      */
-    public function build($data, $deleteSource)
+    public function buildPackage(array $data, bool $deleteSource): void
     {
-        if ($this->packageExists($data['Name'])) {
-            throw new RuntimeException("Package '$data[Name]' already exists");
-            return false;
-        }
-
-        // Set package ID/timestamp
-        $data['Id'] = $this->_serviceManager->get('Library\Now')->getTimestamp();
-
-        $packages = $this->_serviceManager->get('Database\Table\Packages');
-        $storage = $this->_serviceManager->get('Model\Package\Storage\Direct');
-        try {
-            // Obtain archive file
-            $path = $storage->prepare($data);
-            $file = $this->autoArchive($data, $path, $deleteSource);
-            $archiveCreated = ($file != $data['FileLocation']);
-
-            // Determine file size and hash if available
-            if ($data['FileLocation']) {
-                $fileSize = @filesize($file);
-                if (!$fileSize) {
-                    throw new \RuntimeException("Could not determine size of '$file'");
-                }
-                $hash = @sha1_file($file);
-                if (!$hash) {
-                    throw new \RuntimeException("Could not compute SHA1 hash of '$file'");
-                }
-            } else {
-                // No file
-                $fileSize = 0;
-                $hash = null;
-            }
-            $data['Hash'] = $hash;
-            $data['Size'] = $fileSize;
-
-            // Write storage specific data
-            $data['NumFragments'] = $storage->write($data, $file, $deleteSource || $archiveCreated);
-
-            // Create database entries
-            $insert = $packages->getHydrator()->extract(new \ArrayObject($data));
-            if (!$insert['osname']) {
-                throw new \InvalidArgumentException('Invalid platform: ' . $data['Platform']);
-            }
-            $packages->insert($insert);
-        } catch (\Exception $e) {
-            try {
-                $this->delete($data['Name']);
-            } catch (\Exception $e2) {
-                // Ignore error (package does probably not exist at this point)
-                // and return original exception instead
-            }
-            throw new RuntimeException($e->getMessage(), (integer) $e->getCode(), $e);
-        }
-    }
-
-    /**
-     * Create a platform-specific archive if the source file is not already an
-     * archive of the correct type
-     *
-     * This is currently only supported for the 'windows' platform which expects
-     * a ZIP archive. If the Zip extension is not available, the source file is
-     * assumed to be a ZIP archive and a warning is generated.
-     *
-     * The return value is the path to the archive file - either the source file
-     * or a generated archive.
-     *
-     * @param array $data Package data
-     * @param string $path Path where a new archive will be created
-     * @param bool $deleteSource Delete source file after successfully creating an archive
-     * @return string Path to archive file
-     * @throws RuntimeException if an error occurs
-     */
-    public function autoArchive($data, $path, $deleteSource)
-    {
-        $source = $data['FileLocation'];
-        if (!$source) {
-            return $source;
-        }
-
-        $archiveManager = $this->_serviceManager->get('Library\ArchiveManager');
-        switch ($data['Platform']) {
-            case 'windows':
-                $type = \Library\ArchiveManager::ZIP;
-                break;
-            default:
-                // other platforms not implemented yet
-                return $source;
-        }
-        if (!$archiveManager->isSupported($type)) {
-            trigger_error("Support for archive type '$type' not available. Assuming archive.", E_USER_NOTICE);
-            return $source;
-        }
-        if ($archiveManager->isArchive($type, $source)) {
-            // Already an archive of reqired type. Do nothing.
-            return $source;
-        }
-
-        try {
-            $filename = "$path/archive";
-            $archive = $archiveManager->createArchive($type, $filename);
-            $archiveManager->addFile($archive, $source, $data['FileName']);
-            $archiveManager->closeArchive($archive);
-            if ($deleteSource) {
-                \Library\FileObject::unlink($source);
-            }
-        } catch (\Exception $e) {
-            if (isset($archive)) {
-                $archiveManager->closeArchive($archive, true);
-                \Library\FileObject::unlink($filename);
-            }
-            throw new RuntimeException($e->getMessage(), (integer) $e->getCode(), $e);
-        }
-        return $filename;
+        $this->_serviceManager->get(PackageBuilder::class)->buildPackage($data, $deleteSource);
     }
 
     /**
@@ -290,7 +179,7 @@ class PackageManager
      * @param string $name Package name
      * @throws RuntimeException if an error occurs
      */
-    public function delete($name)
+    public function deletePackage($name)
     {
         $packages = $this->_serviceManager->get('Database\Table\Packages');
         $clientConfig = $this->_serviceManager->get('Database\Table\ClientConfig');
@@ -312,22 +201,65 @@ class PackageManager
             $packages->delete(array('fileid' => $id));
             $storage->cleanup($id);
         } catch (\Exception $e) {
-            throw new RuntimeException($e->getMessage(), (integer) $e->getCode(), $e);
+            throw new RuntimeException($e->getMessage(), (int) $e->getCode(), $e);
         }
+    }
+
+    /**
+     * Update a package
+     *
+     * Builds a new package from $newPackageData, calls updateAssignments() with
+     * given parameters and deletes the old package. The passed package object
+     * is updated with new package data.
+     *
+     * @param \Model\Package\Package $package package to be updated
+     * @param array $newPackageData new package data
+     * @param bool $deleteSource Delete source file as soon as possible
+     * @param bool $deployPending Update assignments with status 'pending'
+     * @param bool $deployRunning Update assignments with status 'running'
+     * @param bool $deploySuccess Update assignments with status 'success'
+     * @param bool $deployError Update assignments with status 'error'
+     * @param bool $deployGroups Update assignments for groups
+     */
+    public function updatePackage(
+        \Model\Package\Package $package,
+        array $newPackageData,
+        $deleteSource,
+        $deployPending,
+        $deployRunning,
+        $deploySuccess,
+        $deployError,
+        $deployGroups
+    ) {
+        $oldId = $package['Id'];
+        $oldName = $package['Name'];
+
+        $this->buildPackage($newPackageData, $deleteSource);
+        $newPackage = $this->getPackage($newPackageData['Name']);
+        $this->updateAssignments(
+            $oldId,
+            $newPackage['Id'],
+            $deployPending,
+            $deployRunning,
+            $deploySuccess,
+            $deployError,
+            $deployGroups
+        );
+        $this->deletePackage($oldName);
     }
 
     /**
      * Update package assignments
      *
      * Sets a new package on existing assignments. Updated assignments have
-     * their status reset to "not notified" and their options (force, schedule,
+     * their status reset to "pending" and their options (force, schedule,
      * post cmd) removed.
      *
      * @param integer $oldPackageId package to be replaced
      * @param integer $newPackageId new package
-     * @param bool $deployNonnotified Update assignments with status 'not notified'
+     * @param bool $deployPending Update assignments with status 'pending'
+     * @param bool $deployRunning Update assignments with status 'running'
      * @param bool $deploySuccess Update assignments with status 'success'
-     * @param bool $deployNotified Update assignments with status 'notified'
      * @param bool $deployError Update assignments with status 'error'
      * @param bool $deployGroups Update assignments for groups
      * @throws RuntimeException if an error occurs
@@ -335,35 +267,35 @@ class PackageManager
     public function updateAssignments(
         $oldPackageId,
         $newPackageId,
-        $deployNonnotified,
+        $deployPending,
+        $deployRunning,
         $deploySuccess,
-        $deployNotified,
         $deployError,
         $deployGroups
-    )
-    {
-        if (!($deployNonnotified or $deploySuccess or $deployNotified or $deployError or $deployGroups)) {
+    ) {
+        if (!($deployPending or $deployRunning or $deploySuccess or $deployError or $deployGroups)) {
             return; // nothing to do
         }
 
         $clientConfig = $this->_serviceManager->get('Database\Table\ClientConfig');
         $groupInfo = $this->_serviceManager->get('Database\Table\GroupInfo');
 
-        $where = new \Zend\Db\Sql\Where;
+        $where = new \Laminas\Db\Sql\Where();
         $where->equalTo('ivalue', $oldPackageId);
+        $where->equalTo('name', 'DOWNLOAD');
 
         // Additional filters are only necessary if not all conditions are set
-        if (!($deployNonnotified and $deploySuccess and $deployNotified and $deployError and $deployGroups)) {
+        if (!($deployPending and $deployRunning and $deploySuccess and $deployError and $deployGroups)) {
             $groups = $groupInfo->getSql()->select()->columns(array('hardware_id'));
-            $filters = new \Zend\Db\Sql\Where(null, \Zend\Db\Sql\Where::COMBINED_BY_OR);
-            if ($deployNonnotified) {
+            $filters = new \Laminas\Db\Sql\Where(null, \Laminas\Db\Sql\Where::COMBINED_BY_OR);
+            if ($deployPending) {
                 $filters->isNull('tvalue')->and->notIn('hardware_id', $groups);
+            }
+            if ($deployRunning) {
+                $filters->equalTo('tvalue', \Model\Package\Assignment::RUNNING);
             }
             if ($deploySuccess) {
                 $filters->equalTo('tvalue', \Model\Package\Assignment::SUCCESS);
-            }
-            if ($deployNotified) {
-                $filters->equalTo('tvalue', \Model\Package\Assignment::NOTIFIED);
             }
             if ($deployError) {
                 $filters->like('tvalue', \Model\Package\Assignment::ERROR_PREFIX . '%');
@@ -375,31 +307,40 @@ class PackageManager
         }
 
         $now = $this->_serviceManager->get('Library\Now')->format(\Model\Package\Assignment::DATEFORMAT);
-        try{
+        try {
             // Remove DOWNLOAD_* options from updated assignments
             $subquery = $clientConfig->getSql()
                                      ->select()
                                      ->columns(array('hardware_id'))
-                                     ->where(array('name' => 'DOWNLOAD', $where));
-            $delete = new \Zend\Db\Sql\Where;
-            $delete->equalTo('ivalue', $oldPackageId)
-                   ->in('hardware_id', $subquery)
-                   ->notEqualTo('name', 'DOWNLOAD_SWITCH')
-                   ->like('name', 'DOWNLOAD_%');
-
-            $clientConfig->delete($delete);
+                                     ->where($where);
+            // @codeCoverageIgnoreStart
+            if ($clientConfig->getAdapter()->getPlatform()->getName() == 'MySQL') {
+                // MySQL does not allow subquery on the same table for DELETE
+                // statements. Fetch result as a list instead.
+                $subquery = array_column($clientConfig->selectWith($subquery)->toArray(), 'hardware_id');
+            }
+            // @codeCoverageIgnoreEnd
+            if ($subquery) {
+                // $subquery is either an SQL statement or a non-empty array.
+                $delete = new \Laminas\Db\Sql\Where();
+                $delete->equalTo('ivalue', $oldPackageId)
+                       ->notEqualTo('name', 'DOWNLOAD_SWITCH')
+                       ->in('hardware_id', $subquery)
+                       ->like('name', 'DOWNLOAD_%');
+                $clientConfig->delete($delete);
+            }
 
             // Update package ID and reset status
             $clientConfig->update(
                 array(
                     'ivalue' => $newPackageId,
-                    'tvalue' => \Model\Package\Assignment::NOT_NOTIFIED,
+                    'tvalue' => \Model\Package\Assignment::PENDING,
                     'comments' => $now,
                 ),
-                array('name' => 'DOWNLOAD', $where)
+                $where
             );
         } catch (\Exception $e) {
-            throw new RuntimeException($e->getMessage(), (integer) $e->getCode(), $e);
+            throw new RuntimeException($e->getMessage(), (int) $e->getCode(), $e);
         }
     }
 }
